Colocando o int do aluno para long e dando inicio ao teste Emprestimo

+++ src/main/java/br/com/bibliotecaltv/controller/javabeans/Aluno.java
@@ -26,7 +26,7 @@ public class Aluno {
 	private String nome;
 	
 	@Column(name="numero", nullable = false)
-	private int numero;
+	private Long numero;
 	
 	@Column(name="matricula", nullable = false)
 	private Long matricula;
@@ -53,10 +53,10 @@ public class Aluno {
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
-	public int getNumero() {
+	public Long getNumero() {
 		return numero;
 	}
-	public void setNumero(int numero) {
+	public void setNumero(Long numero) {
 		this.numero = numero;
 	}
 	public Long getMatricula() {
